+++ src/app/interfaces/category.ts
@@ -2,6 +2,7 @@ export interface Category {
   id: number;
   party?: string;
   name: string;
+  image: string;
   /**
    * Extra properties
    */